+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Specification/Domain/Model/ProductValuesCollectionSpec.php
@@ -104,10 +104,10 @@ class ProductValuesCollectionSpec extends ObjectBehavior
         Assert::eq($allTextValues, [$localizableTextareaValues1, $localizableTextareaValues2]);
     }
 
-    public function it_returns_the_values_for_the_main_title()
+    public function it_returns_the_values_for_the_localizable_main_title()
     {
         $commonTextAttribute = $this->givenALocalizableAttributeOfTypeText('a_text');
-        $mainTitleAttribute = $this->givenAMainTitleAttribute('main_title');
+        $mainTitleAttribute = $this->givenALocalizableMainTitleAttribute('main_title');
 
         $commonTextValues = $this->givenRandomValuesForAttribute($commonTextAttribute);
         $mainTitleValues = $this->givenRandomValuesForAttribute($mainTitleAttribute);
@@ -115,7 +115,7 @@ class ProductValuesCollectionSpec extends ObjectBehavior
         $this->add($commonTextValues);
         $this->add($mainTitleValues);
 
-        $this->getMainTitleValues()->shouldReturn($mainTitleValues);
+        $this->getLocalizableMainTitleValues()->shouldReturn($mainTitleValues);
     }
 
     public function it_returns_null_if_there_is_no_main_title()
@@ -125,7 +125,17 @@ class ProductValuesCollectionSpec extends ObjectBehavior
 
         $this->add($commonTextValues);
 
-        $this->getMainTitleValues()->shouldReturn(null);
+        $this->getLocalizableMainTitleValues()->shouldReturn(null);
+    }
+
+    public function it_returns_null_if_the_main_title_is_not_localizable()
+    {
+        $notLocalizableMainTitleAttribute = $this->givenANotLocalizableMainTitleAttribute('not_localizable_main_title');
+        $mainTitleValues = $this->givenRandomValuesForAttribute($notLocalizableMainTitleAttribute);
+
+        $this->add($mainTitleValues);
+
+        $this->getLocalizableMainTitleValues()->shouldReturn(null);
     }
 
     private function givenALocalizableAttributeOfTypeText(string $code): Attribute
@@ -148,9 +158,14 @@ class ProductValuesCollectionSpec extends ObjectBehavior
         return new Attribute(new AttributeCode($code), AttributeType::textarea(), false, false);
     }
 
-    private function givenAMainTitleAttribute(string $code): Attribute
+    private function givenALocalizableMainTitleAttribute(string $code): Attribute
     {
         return new Attribute(new AttributeCode($code), AttributeType::text(), true, true);
+    }
+
+    private function givenANotLocalizableMainTitleAttribute(string $code): Attribute
+    {
+        return new Attribute(new AttributeCode($code), AttributeType::text(), false, true);
     }
 
     private function givenRandomValuesForAttribute(Attribute $attribute): ProductValues